+++ internal/Picker/Picker.js
@@ -66,6 +66,7 @@ const selectDecIcon = selectIcon('decrementIcon', 'arrowlargedown', 'arrowlargel
 const forwardBlur = forward('onBlur'),
 	forwardFocus = forward('onFocus'),
 	forwardKeyDown = forward('onKeyDown'),
+	forwardMouseDown = forward('onMouseDown'),
 	forwardKeyUp = forward('onKeyUp'),
 	forwardWheel = forward('onWheel');
 
@@ -537,6 +538,15 @@ const PickerBase = class extends React.Component {
 
 	emulateMouseUp = new Job(this.clearPressedState, 175)
 
+	handleMouseDown = (ev) => {
+		const {joined, orientation} = this.props;
+		forwardMouseDown(ev, this.props);
+
+		if (joined && orientation === 'horizontal') {
+			this.setIncPickerButtonPressed();
+		}
+	}
+
 	handleUp = () => {
 		if (this.props.joined && (this.pickerButtonPressed !== 0 || this.state.pressed !== 0)) {
 			this.emulateMouseUp.start();
@@ -920,7 +930,7 @@ const PickerBase = class extends React.Component {
 				onKeyDown={this.handleKeyDown}
 				onKeyUp={this.handleKeyUp}
 				onUp={this.handleUp}
-				onMouseDown={isHorizontalJoined ? this.setIncPickerButtonPressed : null}
+				onMouseDown={this.handleMouseDown}
 				onMouseLeave={this.clearPressedState}
 				ref={this.initContainerRef}
 				{...spottablePickerProps}